move streamScopeFieldLabel to common
