+++ packages/studio/src/components/common/StreamScopeField.tsx
@@ -1,6 +1,6 @@
 import PillPickerInput from "../PillPicker/PillPickerInput";
 import { useCallback } from "react";
-import StreamScopeFieldLabel from "../AddPageButton/StreamScopeFieldLabel";
+import StreamScopeFieldLabel from "./StreamScopeFieldLabel";
 import { updateScopeField } from "../StreamScopePicker";
 
 // TODO (SLAP-2918): Combine this with streamScopeFormData once this component
