+++ packages/studio/src/components/common/StreamScopeFieldLabel.tsx
